requirements.txt
uvicorn
fastapi
ollama
google-genai
pydantic
python-dotenv